requirements.txt
Flask==3.0.3
Pillow==11.0.0
segno==1.6.1
